structured api
