+++ api/helper/constats.js
@@ -1,0 +1,7 @@
+export const INTERVAL_ALL = 0;
+
+// blocks
+export const INTERVAL_DAY = 41142;
+export const INTERVAL_WEEK = 287994;
+export const INTERVAL_MONTH = 1234260;
+export const INTERVAL_YEAR = 15016830;
