+++ fetcher.js
@@ -1,248 +1,41 @@
-const apollo = require("apollo-fetch");
-const ethers = require("ethers");
+import { getAlchemicaTotalSupplyDiff } from "./api/alchemica";
+import { INTERVAL_ALL, INTERVAL_DAY } from "./api/helper/constats";
 
-let provider = ethers.getDefaultProvider("http://polygon-rpc.com");
-
-export const INTERVAL_ALL = 0;
-
-// blocks
-export const INTERVAL_DAY = 41142;
-export const INTERVAL_WEEK = 287994;
-export const INTERVAL_MONTH = 1234260;
-export const INTERVAL_YEAR = 15016830;
-
-const gotchiverseSubgraph = apollo.createApolloFetch({
-  uri: "http://157.90.182.138:8000/subgraphs/name/aavegotchi/gotchiverse",
-  //uri: "https://api.thegraph.com/subgraphs/name/aavegotchi/gotchiverse-matic",
-});
-
-const coreMaticSubgraph = apollo.createApolloFetch({
-  //uri: "https://api.thegraph.com/subgraphs/name/aavegotchi/aavegotchi-core-matic",
-  uri: "https://api.thegraph.com/subgraphs/id/QmdLiG6MpgHmGUrUAbG7MgHJRfcusX1P3ucJdKCm6nunS1",
-});
-
-const alchemicaSubgraph = apollo.createApolloFetch({
-  uri: "https://api.thegraph.com/subgraphs/name/aavegotchi/aavegotchi-alchemica",
-});
-
-const aavegotchiSvgSubgraph = apollo.createApolloFetch({
-  uri: "https://api.thegraph.com/subgraphs/name/aavegotchi/aavegotchi-svg",
-});
-
-const gltrStakingSubgraph = apollo.createApolloFetch({
-  uri: "https://api.thegraph.com/subgraphs/name/aavegotchi/aavegotchi-gltr-staking",
-});
-
-export const getCurrentBlock = async () => {
-  const block = await provider.getBlockNumber();
-  return block;
-};
-
-export const getAlchemicaTotalSupplyDiff = async (interval = INTERVAL_ALL) => {
-  let currentBlockNr = await getCurrentBlock();
-  let previousBlockNr = currentBlockNr - interval;
-
-  if (currentBlockNr === previousBlockNr) {
-    return getAlchemicaTotalSupplyFromBlock();
+export default async function Fetcher(url) {
+  let urlParts = url.split("/");
+  console.log(url, urlParts);
+  if (urlParts[1] !== "api") {
+    return false;
   }
 
-  let values = await Promise.all([
-    getAlchemicaTotalSupplyFromBlock(),
-    getAlchemicaTotalSupplyFromBlock(previousBlockNr),
-  ]);
-
-  let diffValues = values[0].map((e, i) => ({
-    name: e.name,
-    totalSupply: e.totalSupply - values[1][i].totalSupply,
-  }));
-
-  return diffValues;
-};
-
-export const getAlchemicaTotalSupplyFromBlock = async (block = 0) => {
-  let query = `{erc20Contracts ${
-    block !== 0 ? `(block: { number: ${block}})` : ""
-  } {
-      name,
-      totalSupply {
-        id
-        value
-      }
-    }}`;
-  const result = await alchemicaSubgraph({ query });
-  const supplies = result.data.erc20Contracts.map((e) => ({
-    name: e.name,
-    totalSupply: e.totalSupply.value,
-  }));
-
-  return supplies;
-};
-
-export const getPoolInfo = async (interval = INTERVAL_ALL) => {
-  let query = `{
-    erc20Balances(
-      where: {account_in: ["0x1fe64677ab1397e20a1211afae2758570fea1b8c"], contract_not: "0x3801c3b3b5c98f88a9c9005966aa96aa440b9afc"}
-    ) {
-      value
-      account {
-        id
-      }
-      contract {
-        name
-        id
-        totalSupply {
-          value
-        }
-      }
-    }
-  }`;
-
-  const result = await gltrStakingSubgraph({ query });
-  const balancesOfPools = result.data.erc20Balances;
-
-  return balancesOfPools.map((e) => ({
-    pool: e.contract.id,
-    staked: e.value,
-    totalSupply: e.contract.totalSupply.value,
-    percentageStaked: (e.value / e.contract.totalSupply.value) * 100,
-  }));
-};
-
-export const getStats = async () => {
-  // let query = `{stat(id:"overall") {
-  //     countChannelAlchemicaEvents
-  //     countParcelInstallations
-  //     countInstallationTypes
-  //     countUpgradesInitiated
-  //     alchemicaSpendOnInstallations
-  //     alchemicaSpendOnUpgrades
-  //     alchemicaSpendOnTiles
-  //     alchemicaSpendTotal
-  //     alchemicaChanneledTotal
-  //     alchemicaClaimedTotal
-  //     alchemicaExitedTotal
-  //     tilesMinted
-  //     installationsMintedTotal
-  //     installationsUpgradedTotal
-  //   }}`;
-
-  // const result = await gotchiverseSubgraph({ query });
-
-  // console.log(JSON.stringify(result));
-  // return result.data.stat;
-
-  let data = {
-    countChannelAlchemicaEvents: "261158",
-    countParcelInstallations: "16656",
-    countInstallationTypes: "0",
-    countUpgradesInitiated: "0",
-    alchemicaSpendOnInstallations: [
-      "9325100000000000000000000",
-      "5018550000000000000000000",
-      "4073025000000000000000000",
-      "1218920000000000000000000",
-    ],
-    alchemicaSpendOnUpgrades: [
-      "2224400000000000000000000",
-      "1150950000000000000000000",
-      "784225000000000000000000",
-      "193030000000000000000000",
-    ],
-    alchemicaSpendOnTiles: [
-      "389300000000000000000000",
-      "389300000000000000000000",
-      "1167900000000000000000000",
-      "389300000000000000000000",
-    ],
-    alchemicaSpendTotal: [
-      "11938800000000000000000000",
-      "6558800000000000000000000",
-      "6025150000000000000000000",
-      "1801250000000000000000000",
-    ],
-    alchemicaChanneledTotal: [
-      "32794492800000000000000000",
-      "16397246400000000000000000",
-      "8198623200000000000000000",
-      "3279449280000000000000000",
-    ],
-    alchemicaClaimedTotal: ["0", "0", "0", "0"],
-    alchemicaExitedTotal: ["0", "0", "0", "0"],
-    tilesMinted: "15572",
-    installationsMintedTotal: "53406",
-    installationsUpgradedTotal: "24798",
-  };
-
-  return data;
-};
-
-export const getGotchis = async (interval = INTERVAL_ALL) => {
-  let query = `{statistics {
-      aavegotchisClaimed 
-      aavegotchisSacrificed
-      aavegotchisBorrowed
-    }}`;
-
-  let result = await coreMaticSubgraph({ query });
-  let gotchisClaimed = result.data.statistics[0].aavegotchisClaimed;
-
-  query = `{user(id:"0x0000000000000000000000000000000000000000") {
-      gotchisOwned(first: 1000) {
-        id
-      }
-    }}`;
-
-  result = await coreMaticSubgraph({ query });
-  let gotchisSacrificed = result.data.user.gotchisOwned.length.toString();
-
-  return { gotchisClaimed, gotchisSacrificed };
-};
-
-export const getActiveWallets = async (interval = INTERVAL_ALL) => {
-  const data = await Promise.all([
-    fetch(
-      "https://dappradar.com/v2/api/dapp/polygon/games/aavegotchi/statistic/24h?currency=USD"
-    ),
-    fetch(
-      "https://dappradar.com/v2/api/dapp/polygon/games/aavegotchi/statistic/week?currency=USD"
-    ),
-    fetch(
-      "https://dappradar.com/v2/api/dapp/polygon/games/aavegotchi/statistic/month?currency=USD"
-    ),
-  ]);
-
-  for (let i = 0; i < data.length; i++) {
-    data[i] = (await data[i].json()).userActivity.toString();
+  if (!urlParts[2].includes(["alchemica"])) {
+    return false;
   }
 
-  return data;
-};
+  let fetchMethod = getCategoryMethod(urlParts[2]);
+  if (!fetchMethod) {
+    return false;
+  }
 
-export const getBurnedGLTR = async () => {
-  // let query = `{stat(id:"overall") {
-  //   gltrSpendTotal
-  //   gltrSpendOnCrafts
-  //   gltrSpendOnUpgrades
-  // }}`;
+  let interval = INTERVAL_ALL;
+  if (urlParts[3]) {
+    interval = INTERVAL_DAY * parseInt(urlParts[3].substring(-1));
+  }
 
-  // const { data } = await gotchiverseSubgraph({ query });
-  // const stat = data.stat;
+  let asTimeSeries = false;
+  if (urlParts[4] == "series") {
+    asTimeSeries = true;
+  }
 
-  // Object.keys(stat).forEach((e) => {
-  //   let value = stat[e];
-  //   stat[e] =
-  //     value.length >= 18
-  //       ? value.slice(0, -18) + "." + value.slice(-18, -16)
-  //       : value;
-  // });
+  const result = await fetchMethod(interval, asTimeSeries);
+  console.log(result);
+  return {};
+}
 
-  // return stat;
+function getCategoryMethod(category) {
+  if (category == "alchemica") {
+    return getAlchemicaTotalSupplyDiff;
+  }
 
-  let data = {
-    gltrSpendTotal: "11000",
-    gltrSpendOnCrafts: "1000",
-    gltrSpendOnUpgrades: "10000",
-  };
-
-  return data;
-};
+  return false;
+}
